pedEbear/pedEbear/requirements.txt
beautifulsoup4
requests
pytesseract
Pillow
tweepy==4.13.0
python-dotenv